requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
asyncpg==0.29.0
asyncio==3.4.3
pydantic==2.8.2
bcrypt==4.1.3
python-dotenv==1.0.1
pillow==10.4.0
python-jose==3.3.0
aiofiles==24.1.0